CCKEY-260 Fixed browser password behavior by adding HTML and CSS properties
Browser autosaving is always enabled by default and only can be
addressed by a developer if the input field is not recognized by the
browser.
Chrome does not follow HTML standards and ignores an implemented
autocomplete="off". Therefore a CSS class had to be used which styles
text into password-bullets.
Further information regarding chrome can be found in this issue:
https://bugs.chromium.org/p/chromium/issues/detail?id=468153